separated snake and node
